app/requirements.txt
asgiref==3.4.1
astroid==2.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==4.0
django-admin-interface==0.18.3
django-axes==5.28.0
django-ckeditor==6.2.0
django-colorfield==0.5.0
django-cors-middleware==1.5.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-ipware==4.0.2
django-js-asset==1.2.2
django-split-settings==1.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
